XMC flash output drive initial attemt
adds flash_xmc_check() and some functions for sending custom
spi commands through spi0.

These functions operate the registers in the same way that
  esptool.py write_flash_status --bytes 3 0x600000
does. The register is set, and all is well.

The only problem left is reading the register in this code is not
working yet. Once that's done I will produce a stripped down version
to put in the Arduino SDK startup.