Fix reading the register.
spi0_command() can now be used to send custom SPI commands on SPI0 (the interface
connected to the flash chip).

Note that use of the specific ROM calls for most operations is preferred
as they make optimal use of the SPI controller, respect the current Flash
IO mode and do not occupy precious IRAM.

spi0_command() is only preferred where the ROM can not perform an operation (eg accessing
flash Status Register 3) as it uses single bit serial access.

Unfortunately spi0_command() must be ICACHE_RAM_ATTR at a cost of 392 bytes because
if it isn't a page load from flash can be triggered while we are configuring the
SPI interface.