Add PRECACHE_ macros and automatic size calculation
Now that the precaching is done within the function, we can use some labels
and macros to calculate the size of the code in bytes at compile time.

Preloading code into the cache can now be done reliably on small blocks of
code within a function (or whole functions) with three simple macros which
make use of a small function.